Finshed my second homework
It took me less time than the first one..
What helped me is when we made the prototype on the white board how to devide the blocks.
Also, when we did the warm up.
I faced an obstacle when I was trying to centering one of the children excepting the others